requirements.txt
python-telegram-bot==20.9
pandas==2.3.1
openpyxl==3.1.2
python-dotenv==1.0.0